Time: 6 ms (74.64%), Space: 22.9 MB (54.8%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -5,7 +5,7 @@ public:
         int end = *max_element(piles.begin(), piles.end());
 
         while(st<=end){
-            long int takenHours=0;
+         long long takenHours=0;
             int mid= st+(end-st)/2;
 
             for(int i=0;i<piles.size();i++){
